Deposit. Changed the input type for all account fields (Checking, Saving, Credit, Loan). from float to text.
All values from input will be saved as a string.

I modified the names of some of the input fields so they are not the same but remain using the same naming convention.

I added IDs to the table for the description of the checking row.

I added the remove functions for deposit, withdrawal and transfer forms so I can hide the forms after I am finished with them.
I added notes to organized my code.

I used the Number() to capture the values for the checkingBalance and the depositAmt so I can add them and get the newBalance.
I verified this by using a console.table.

I modified the newBalance value in the users array so I can display the dashboard with the current balance.

NEXT OBJECTIVE: Savings modification.